I defined the sum variable to float for showing the float amount in average

+++ week01/Exercise4/Program.cs
@@ -9,7 +9,7 @@ class Program
         List<int> numbers = new List<int>();
         Console.WriteLine("Enter a list of numbers, type 0 when finished.");
         int number;
-        int sumNumbers = 0; 
+        float sumNumbers = 0; 
         int size = 0;
         float average = 0;
         int largestNumber = 0;
